feat(renovate-runner): Renovate docker image (#23)
The default docker image for renovate uses an
old version of yarn which is causing a ton of
churn in the tarn.lock file.  This gives us control

Additionally, the dependencies that we use within the Dockerfiles have been added to the package.json files so that renovate will let us know when they need to be updated (and there will be a strong clue which images need to be updated).